# packages/pth-console/operator-console-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTL Operator Console API",
    "version": "1.0.0",
    "description": "Loopback-only Console HTTP API. See docs/pth/console-protocol.md for security rules."
  },
  "servers": [{ "url": "http://127.0.0.1:3194" }],
  "paths": {
...
    "/api/v1/debug/workers": {
      "get": {
        "summary": "Worker inspection projection",
        "operationId": "getDebugWorkers",
        "responses": {
          "200": {
            "description": "worker list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["workers", "tenant", "space"],
                  "properties": {
                    "workers": { "type": "array", "items": { "$ref": "#/components/schemas/WorkerRow" } },
                    "tenant": { "type": "string" },
                    "space": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" },
          "503": { "$ref": "#/components/responses/Error503" }
        }
      }
    },
    "/api/v1/memory/summary": {
      "get": {
        "summary": "Memory summary by canonical type",
        "operationId": "getMemorySummary",
        "responses": {
          "200": {
            "description": "summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["byType"],
                  "properties": {
                    "byType": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "required": ["count", "bytes"],
                        "properties": { "count": { "type": "integer" }, "bytes": { "type": "integer" } }
                      }
                    },
                    "totals": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/memory/entries": {
      "get": {
        "summary": "Memory entries page",
        "operationId": "listMemoryEntries",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "anchor", "in": "query", "schema": { "type": "string" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items", "cursor", "total"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/MemoryItem" } },
                    "cursor": { "type": "string", "nullable": true },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/memory/entries/{id}": {
      "get": {
        "summary": "Memory entry detail",
        "operationId": "getMemoryEntry",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "item",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryItem" } } }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/memory/entries/{id}/revisions": {
      "get": {
        "summary": "Memory revision history",
        "operationId": "getMemoryRevisions",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "revisions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["revision", "action", "time", "type"],
                    "properties": {
                      "revision": { "type": "integer", "nullable": true },
                      "action": { "type": "string" },
                      "time": { "type": "string" },
                      "type": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/config/ptl": {
      "get": {
        "summary": "PTL local config descriptor",
        "operationId": "getPtlConfig",
        "responses": {
          "200": {
            "description": "config items",
            "content": {
              "application/json": {
                "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/ConfigItem" } } } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" }
        }
      }
    },
    "/api/v1/config/pth": {
      "get": {
        "summary": "PTH config descriptor (redacted)",
        "operationId": "getPthConfig",
        "responses": {
          "200": {
            "description": "config items",
            "content": {
              "application/json": {
                "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/ConfigItem" } } } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/roles": {
      "get": {
        "summary": "Runtime catalog roles",
        "operationId": "getRoles",
        "responses": {
          "200": {
            "description": "role items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "revision"],
                        "properties": {
                          "id": { "type": "string" },
                          "parent": { "type": "string", "nullable": true },
                          "revision": { "type": "string" },
                          "family": { "type": "string" },
                          "tags": { "type": "array", "items": { "type": "string" } },
                          "capabilities": { "type": "array", "items": { "type": "string" } }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "WorkerRow": {
        "type": "object",
        "required": ["workerId", "batchId", "roleId", "roleRevision", "lifecycle", "workMode", "regions", "workingSet", "toolNames", "skillIds"],
        "properties": {
          "workerId": { "type": "string" },
          "batchId": { "type": "string" },
          "roleId": { "type": "string" },
          "roleRevision": { "type": "string" },
          "lifecycle": { "type": "string" },
          "workMode": { "type": "string", "nullable": true },
          "taskId": { "type": "string", "nullable": true },
          "leaseId": { "type": "string", "nullable": true },
          "heartbeatAt": { "type": "string", "nullable": true },
          "regions": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["regionId"],
              "properties": { "regionId": { "type": "string" }, "weights": { "type": "number", "nullable": true } }
            }
          },
          "workingSet": { "type": "array", "items": { "type": "string" } },
          "toolNames": { "type": "array", "items": { "type": "string" } },
          "skillIds": { "type": "array", "items": { "type": "string" } }
        }
      },
      "MemoryItem": {
        "type": "object",
        "required": ["id", "kind", "status"],
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "nullable": true },
          "kind": { "type": "string" },
          "status": { "type": "string" },
          "anchors": { "type": "array", "items": { "type": "string" } },
          "version": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "nullable": true },
          "updatedAt": { "type": "string", "nullable": true },
          "contentBytes": { "type": "integer", "nullable": true }
        }
      },
      "ConfigItem": {
        "type": "object",
        "required": ["key", "group", "type", "scope", "source", "secret"],
        "properties": {
          "key": { "type": "string" },
          "group": { "type": "string" },
          "type": { "type": "string" },
          "scope": { "type": "string" },
          "source": { "type": "string" },
          "runtimeMutable": { "type": "boolean" },
          "restartRequired": { "type": "boolean" },
          "description": { "type": "string" },
          "secret": { "type": "boolean" },
          "defaultValue": {},
          "effectiveValue": {},
          "sourceDetail": { "type": "string", "nullable": true }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_-]{0,63}$" },
              "message": { "type": "string" },
              "requestId": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
...
      "Error400": { "description": "bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error401": { "description": "unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Error502": { "description": "upstream failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error503": { "description": "unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}